fix a typo and move things around
